Cleanup and Formating 	Using https://raw.github.com/android/platform_development/master/ide/eclipse/android-formatting.xml 	Remove unused imports 	Remove warnings

+++ src/com/motorola/fmradio/FMUtil.java
@@ -1,14 +1,11 @@
 package com.motorola.fmradio;
 
-import android.content.ContentResolver;
 import android.media.AudioManager;
 
 import android.content.Context;
 import android.content.res.Resources;
 import android.database.Cursor;
-import android.net.Uri;
 import android.text.TextUtils;
-import android.view.Gravity;
 import android.widget.Toast;
 
 import java.text.DecimalFormat;
@@ -23,14 +20,13 @@ public class FMUtil {
     // public static final int STREAM_FM = AudioManager.STREAM_FM;
     public static final int STREAM_FM = AudioManager.STREAM_MUSIC;
 
-    public static final String[] PROJECTION = new String[] {
-        FMDataProvider.Channels.ID, FMDataProvider.Channels.FREQUENCY,
-        FMDataProvider.Channels.NAME, FMDataProvider.Channels.RDS_NAME
-    };
-    public static  final int CHANNEL_COLUMN_ID = 0;
-    public static  final int CHANNEL_COLUMN_FREQ = 1;
-    public static  final int CHANNEL_COLUMN_NAME = 2;
-    public static  final int CHANNEL_COLUMN_RDSNAME = 3;
+    public static final String[] PROJECTION = new String[] { FMDataProvider.Channels.ID,
+            FMDataProvider.Channels.FREQUENCY, FMDataProvider.Channels.NAME,
+            FMDataProvider.Channels.RDS_NAME };
+    public static final int CHANNEL_COLUMN_ID = 0;
+    public static final int CHANNEL_COLUMN_FREQ = 1;
+    public static final int CHANNEL_COLUMN_NAME = 2;
+    public static final int CHANNEL_COLUMN_RDSNAME = 3;
 
     public static final int FREQ_RATE = 1000;
 
@@ -90,7 +86,7 @@ public class FMUtil {
     }
 
     public static String formatFrequency(Context context, int frequency) {
-        float freq = (float) frequency / 1000.0F;
+        float freq = frequency / 1000.0F;
         DecimalFormat formatter = new DecimalFormat(".0");
         return formatter.format(freq) + context.getString(R.string.mhz);
     }
